Update feature_key_overrides.md
The modifier example lacks the bit in which the key override is globally defined, which all the other examples have. It took me forever to realize that this missing bit of code was the reason it would not work. Furthermore, the text mentions different mods than used in the example and some more explanations don't hurt.